components/ingest-gateway/requirements.txt
# WIP Ingest Gateway Service Dependencies (Development)

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Message queue
nats-py>=2.6.0

# HTTP client (for REST API calls)
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Utilities
python-json-logger>=2.0.0
